+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - `Innmind\IO\Sockets::servers()`
 - `Innmind\IO\Sockets\Servers`
+- `Innmind\IO\Sockets\Server\Pool`
 
 ## 2.6.0 - 2024-03-09